// sitat-fe-workspace/projects/app-contratti/src/app/modules/models/fasi/fase-aggiudicazione-semp.model.ts
export interface FaseAggiudicazioneSempEntry {
    altreSomme?: number;
    astaElettronica?: string;
    codGara?: number;
    codLotto?: number;
    counter?: number;
    dataAtto?: Date;
    dataStipula?: Date;
    dataVerbAgg?: Date;
    durataCon?: number;
    importoAgg?: number;
    importoComplAppalto?: number;
    importoComplIntervento?: number;
    importoDisposizione?: number;
    importoIva?: number;
    importoSicurezza?: number;
    importoSub?: number;
    iva?: number;
    numeroAtto?: string;
    percOffAumento?: number;
    percRibassoAgg?: number;
    pubblicata?: boolean;
    tipoAtto?: number;
}

export interface FaseAggSempInsertForm {
    altreSomme?: number;
    astaElettronica?: string;
    codGara?: number;
    codLotto?: number;
    counter?: number;
    dataAtto?: Date;
    dataStipula?: Date;
    dataVerbAgg?: Date;
    durataCon?: number;
    importoAgg?: number;
    importoComplAppalto?: number;
    importoComplIntervento?: number;
    importoDisposizione?: number;
    importoIva?: number;
    importoSicurezza?: number;
    importoSub?: number;
    iva?: number;
    num?: number;
    numeroAtto?: string;
    percOffAumento?: number;
    percRibassoAgg?: number;
    tipoAtto?: number;
}